+++ yane-core/src/main/java/de/turing85/yane/impl/cpu6502/AddressingResult.java
@@ -6,6 +6,6 @@ import lombok.*;
 @AllArgsConstructor
 public class AddressingResult {
   Register register;
-  byte valueRead;
+  int valueRead;
   int additionalCyclesNeeded;
 }
